Compilation fixes added input_rec node based on input node
