+++ src/nodes/force_fps.cpp
@@ -1,6 +1,6 @@
 #include "node_common.hpp"
 
-template<typename T> class ForceFPS: public NonBlockingNode<ForceFPS<T>>, public NodeSISO<T, T>, public IFrameRateSource, public ITimeBaseSource {
+template<typename T> class ForceFPS: public NodeSISO<T, T>, public IFrameRateSource, public ITimeBaseSource {
 private:
     av::Rational fps_;
     av::Timestamp frame_delta_;
@@ -32,18 +32,27 @@ public:
         }
         logstream << "Set timebase " << timebase_ << ", frame rate " << fps_ << ", frame delta " << frame_delta_;
     }
-    virtual void processNonBlocking(EventLoop& evl, bool ticks) override {
-        bool process_next;
-        do {
-            process_next = false;
-            T* ptr = this->source_->peek(0);
-            if (ptr==nullptr) {
-                // no data available in queue
-                if (!ticks) {
-                    // retry when we have packet in source queue
-                    this->processWhenSignalled(this->edgeSource()->edge()->producedEvent());
-                }
-                return;
+    virtual void process() {
+        T* ptr = this->source_->peek();
+        if (ptr==nullptr) return;
+        
+        T &pkt = *ptr;
+        if (!pkt.isValid()) return;
+        
+        if (last_printed_stats_.isNoPts()) {
+            last_printed_stats_ = wallclock.ts();
+        }
+        
+        av::Timestamp in_ts = pkt.pts();
+        if ((in_ts.timebase() != timebase_) || (pkt.timeBase() != timebase_)) {
+            pkt.setTimeBase(timebase_);
+            in_ts = rescaleTS(in_ts, timebase_);
+        }
+        if (last_ts_.isValid()) {
+            av::Timestamp delta_from_last = in_ts - last_ts_;
+            bool discontinuity = (delta_from_last.seconds() > 0.5) || (delta_from_last.timestamp() < 0);
+            if (discontinuity) {
+                logstream << "Discontinuity " << last_ts_ << " -> " << in_ts;
             }
             /*if (delta_from_last == frame_delta_) {
                 logstream << "Frame PTS = " << in_ts << " perfectly aligned";
@@ -69,6 +78,28 @@ public:
                         this->sink_->put(last_frame_);
                         next_ts_ = addTS(next_ts_, frame_delta_);
                     }
+                    // now in_ts <= next_ts_
+                    //  if in_ts == next_ts, all OK
+                    //  if in_ts < next_ts_, it means that in_ts is too small and unaligned
+                    //   aligning it is dangerous, so...
+                    //   so continue to  the "PTS too small" logic!
+                    //in_ts = next_ts_; // align
+                    /*if (in_ts != next_ts_) {
+                        logstream << " :/ unaligned";
+                    } else {
+                        logstream << " :) aligned";
+                    }*/
+                }
+                if (in_ts < next_ts_) {
+                    // PTS too small
+                    // input frame rate too high
+                    // save received frame as last_frame_
+                    // and drop it!
+                    setLast(pkt, true);
+                    total_in_++;
+                    this->source_->pop();
+                    //logstream << "Dropping frame PTS = " << in_ts;
+                    return;
                 }
             }
         }
